*** empty log message ***
git-svn-id: https://ncisvn.nci.nih.gov/svn/cagrid/trunk@6668 dd7b0c16-3c94-4492-92de-414b6a06f0b1

+++ cagrid-1-0/caGrid/projects/introduce/src/java/Portal/gov/nih/nci/cagrid/introduce/portal/modification/discovery/files/FileTypesSelectionComponent.java
@@ -6,6 +6,7 @@ import gov.nih.nci.cagrid.introduce.IntroduceConstants;
 import gov.nih.nci.cagrid.introduce.ResourceManager;
 import gov.nih.nci.cagrid.introduce.beans.extension.DiscoveryExtensionDescriptionType;
 import gov.nih.nci.cagrid.introduce.beans.namespace.NamespaceType;
+import gov.nih.nci.cagrid.introduce.beans.namespace.NamespacesType;
 import gov.nih.nci.cagrid.introduce.common.CommonTools;
 import gov.nih.nci.cagrid.introduce.common.FileFilters;
 import gov.nih.nci.cagrid.introduce.portal.extension.ExtensionTools;
@@ -49,8 +50,8 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 	/**
 	 * This method initializes
 	 */
-	public FileTypesSelectionComponent(DiscoveryExtensionDescriptionType descriptor) {
-		super(descriptor);
+	public FileTypesSelectionComponent(DiscoveryExtensionDescriptionType descriptor, NamespacesType types) {
+		super(descriptor, types);
 		initialize();
 	}
 
@@ -80,10 +81,10 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 		gridBagConstraints9.gridy = 1;
 		gridBagConstraints9.insets = new java.awt.Insets(2, 2, 2, 2);
 		gridBagConstraints9.gridx = 0;
-		this.namespaceLabel = new JLabel();
-		this.namespaceLabel.setText("Namespace");
+		namespaceLabel = new JLabel();
+		namespaceLabel.setText("Namespace");
 		this.setLayout(new GridBagLayout());
-		this.add(this.namespaceLabel, gridBagConstraints9);
+		this.add(namespaceLabel, gridBagConstraints9);
 		this.add(getBrowseButton(), gridBagConstraints);
 		this.add(getNamespaceText(), gridBagConstraints1);
 		this.add(getFilenameText(), gridBagConstraints2);
@@ -96,21 +97,19 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 	 * @return javax.swing.JButton
 	 */
 	private JButton getBrowseButton() {
-		if (this.browseButton == null) {
-			this.browseButton = new JButton();
-			this.browseButton.setText("Browse");
-			this.browseButton.addActionListener(new java.awt.event.ActionListener() {
+		if (browseButton == null) {
+			browseButton = new JButton();
+			browseButton.setText("Browse");
+			browseButton.addActionListener(new java.awt.event.ActionListener() {
 				public void actionPerformed(java.awt.event.ActionEvent e) {
 					try {
 						String selectedFilename = ResourceManager.promptFile(null, FileFilters.XSD_FILTER);
 						if (selectedFilename != null) {
-							FileTypesSelectionComponent.this.currentFile = new File(selectedFilename).getAbsolutePath();
-							getFilenameText().setText(FileTypesSelectionComponent.this.currentFile);
-							Document doc = XMLUtilities
-								.fileNameToDocument(FileTypesSelectionComponent.this.currentFile);
-							FileTypesSelectionComponent.this.currentNamespace = new Namespace(doc.getRootElement()
-								.getAttributeValue("targetNamespace"));
-							getNamespaceText().setText(FileTypesSelectionComponent.this.currentNamespace.getRaw());
+							currentFile = new File(selectedFilename).getAbsolutePath();
+							getFilenameText().setText(currentFile);
+							Document doc = XMLUtilities.fileNameToDocument(currentFile);
+							currentNamespace = new Namespace(doc.getRootElement().getAttributeValue("targetNamespace"));
+							getNamespaceText().setText(currentNamespace.getRaw());
 						}
 					} catch (Exception ex) {
 						ErrorDialog.showErrorDialog("Please make sure the file is a valid XML Schema", ex);
@@ -118,27 +117,41 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 				}
 			});
 		}
-		return this.browseButton;
+		return browseButton;
 	}
 
 
-	public NamespaceType[] createNamespaceType(File schemaDestinationDir) {
+	public NamespaceType[] createNamespaceType(File schemaDestinationDir, String namespaceExistsPolicy) {
 		try {
+			if (currentNamespace == null) {
+				addError("Please select a schema file.");
+				return null;
+			}
+			if (namespaceAlreadyExists(currentNamespace.getRaw()) && namespaceExistsPolicy.equals(ERROR)) {
+				addError("Namespace already exists.");
+				return null;
+			}
+
+			boolean result = checkAgainstPolicy(currentFile, new HashSet(), namespaceExistsPolicy);
+			if (result == false) {
+				return null;
+			}
+
 			List namespaces = new ArrayList();
 
-			String currentFileName = (new File(this.currentFile)).getName();
+			String currentFileName = (new File(currentFile)).getName();
 			NamespaceType root = new NamespaceType();
 			// set the package name
-			String packageName = CommonTools.getPackageName(this.currentNamespace);
+			String packageName = CommonTools.getPackageName(currentNamespace);
 			root.setPackageName(packageName);
-			root.setNamespace(this.currentNamespace.getRaw());
+			root.setNamespace(currentNamespace.getRaw());
 			root.setLocation("./" + currentFileName);
 
 			namespaces.add(root);
 
-			ExtensionTools.setSchemaElements(root, XMLUtilities.fileNameToDocument(this.currentFile));
+			ExtensionTools.setSchemaElements(root, XMLUtilities.fileNameToDocument(currentFile));
 			Set storedSchemas = new HashSet();
-			copySchemas(this.currentFile, schemaDestinationDir, new HashSet(), storedSchemas);
+			copySchemas(currentFile, schemaDestinationDir, new HashSet(), storedSchemas, namespaceExistsPolicy);
 			Iterator schemaFileIter = storedSchemas.iterator();
 			while (schemaFileIter.hasNext()) {
 				File storedSchemaFile = new File((String) schemaFileIter.next());
@@ -158,17 +171,80 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 	}
 
 
-	public static void copySchemas(String fileName, File copyToDirectory, Set visitedSchemas, Set storedSchemas)
-		throws Exception {
+	private boolean checkAgainstPolicy(String fileName, Set visitedSchemas, String namespaceExistsPolicy) {
+		try {
+			File schemaFile = new File(fileName);
+			// mark the schema as visited
+			visitedSchemas.add(schemaFile.getCanonicalPath());
+			// look for imports
+			Document schema = XMLUtilities.fileNameToDocument(schemaFile.getCanonicalPath());
+			List importEls = schema.getRootElement().getChildren("import",
+				schema.getRootElement().getNamespace(IntroduceConstants.W3CNAMESPACE));
+			if (importEls != null) {
+				for (int i = 0; i < importEls.size(); i++) {
+					org.jdom.Element importEl = (org.jdom.Element) importEls.get(i);
+					String location = importEl.getAttributeValue("schemaLocation");
+					if (location != null) {
+						String namespace = importEl.getAttributeValue("namespace");
+						if (namespace != null) {
+							// see if namespace already exists if so check
+							// property
+							// to see if supposed to error;
+							if (namespaceAlreadyExists(namespace)) {
+								if (namespaceExistsPolicy.equals(ERROR)) {
+									addError("Imported namespace already exists: "
+										+ namespace
+										+ ". \nIf you want this schema to be overwriten please change the policy to \"ignore\" in the introduce preferences menu");
+									return false;
+								}
+							}
+
+						}
+
+						File currentPath = schemaFile.getCanonicalFile().getParentFile();
+						if (!schemaFile.equals(new File(currentPath.getCanonicalPath() + File.separator + location))) {
+							File importedSchema = new File(currentPath + File.separator + location);
+							if (!visitedSchemas.contains(importedSchema.getCanonicalPath())) {
+								// only copy schemas not yet visited
+								boolean result = checkAgainstPolicy(importedSchema.getCanonicalPath(), visitedSchemas,
+									namespaceExistsPolicy);
+								if (result == false) {
+									return false;
+								}
+							}
+						} else {
+							System.err.println("WARNING: Schema is importing itself. " + schemaFile);
+						}
+
+					}
+				}
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			addError(e.getMessage());
+			return false;
+		}
+		return true;
+	}
+
+
+	public void copySchemas(String fileName, File copyToDirectory, Set visitedSchemas, Set storedSchemas,
+		String namespaceExistsPolicy) throws Exception {
 		File schemaFile = new File(fileName);
-		System.out.println("Copying schema " + fileName + " to " + copyToDirectory.getCanonicalPath());
-		File outFile = new File(copyToDirectory.getCanonicalPath() + File.separator + schemaFile.getName());
-		Utils.copyFile(schemaFile, outFile);
-		storedSchemas.add(outFile.getAbsolutePath());
-		// mark the schema as visited
-		visitedSchemas.add(schemaFile.getCanonicalPath());
-		// look for imports
 		Document schema = XMLUtilities.fileNameToDocument(schemaFile.getCanonicalPath());
+		String namespaceURI = schema.getRootElement().getAttribute("targetNamespace").getValue();
+		if (namespaceAlreadyExists(namespaceURI) && namespaceExistsPolicy.equals(IGNORE)) {
+			// do nothing just ignore.....
+		} else {
+			System.out.println("Copying schema " + fileName + " to " + copyToDirectory.getCanonicalPath());
+			File outFile = new File(copyToDirectory.getCanonicalPath() + File.separator + schemaFile.getName());
+			Utils.copyFile(schemaFile, outFile);
+			storedSchemas.add(outFile.getAbsolutePath());
+			// mark the schema as visited
+			visitedSchemas.add(schemaFile.getCanonicalPath());
+		}
+
+		// look for imports
 		List importEls = schema.getRootElement().getChildren("import",
 			schema.getRootElement().getNamespace(IntroduceConstants.W3CNAMESPACE));
 		if (importEls != null) {
@@ -182,7 +258,8 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 						if (!visitedSchemas.contains(importedSchema.getCanonicalPath())) {
 							// only copy schemas not yet visited
 							copySchemas(importedSchema.getCanonicalPath(), new File(copyToDirectory.getCanonicalFile()
-								+ File.separator + location).getParentFile(), visitedSchemas, storedSchemas);
+								+ File.separator + location).getParentFile(), visitedSchemas, storedSchemas,
+								namespaceExistsPolicy);
 						}
 					} else {
 						System.err.println("WARNING: Schema is importing itself. " + schemaFile);
@@ -199,11 +276,11 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 	 * @return javax.swing.JTextField
 	 */
 	private JTextField getNamespaceText() {
-		if (this.namespaceText == null) {
-			this.namespaceText = new JTextField();
-			this.namespaceText.setEditable(false);
+		if (namespaceText == null) {
+			namespaceText = new JTextField();
+			namespaceText.setEditable(false);
 		}
-		return this.namespaceText;
+		return namespaceText;
 	}
 
 
@@ -213,12 +290,12 @@ public class FileTypesSelectionComponent extends NamespaceTypeDiscoveryComponent
 	 * @return javax.swing.JTextField
 	 */
 	private JTextField getFilenameText() {
-		if (this.filenameText == null) {
-			this.filenameText = new JTextField();
-			this.filenameText.setEnabled(true);
-			this.filenameText.setEditable(false);
+		if (filenameText == null) {
+			filenameText = new JTextField();
+			filenameText.setEnabled(true);
+			filenameText.setEditable(false);
 		}
-		return this.filenameText;
+		return filenameText;
 	}
 
 } // @jve:decl-index=0:visual-constraint="16,10"
